+++ lib/http_spec.ex
@@ -36,4 +36,6 @@ defmodule HTTPSpec do
 
   Check out their docs for more available functionalities.
   """
+
+  @type fields :: {name :: String.t(), value :: String.t()}
 end